+++ src/main/java/T145/magistics/pulses/PulseNEI.java
@@ -2,7 +2,7 @@ package T145.magistics.pulses;
 
 import T145.magistics.blocks.BlockNetherFurnace;
 import T145.magistics.items.ItemDummy;
-import T145.magistics.lib.Injector;
+import T145.magistics.lib.utils.Injector;
 import T145.magistics.pulses.core.ServerPulse;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
