+++ src/main/java/T145/magistics/tiles/TileChestHungryEnder.java
@@ -2,8 +2,8 @@ package T145.magistics.tiles;
 
 import T145.magistics.api.tiles.IFacing;
 import T145.magistics.api.tiles.IOwned;
-import T145.magistics.api.tiles.TileChest;
-import T145.magistics.lib.ModHelper;
+import T145.magistics.lib.utils.PlayerUtils;
+import T145.magistics.tiles.core.TileChest;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.InventoryEnderChest;
 import net.minecraft.item.ItemStack;
@@ -16,7 +16,7 @@ public class TileChestHungryEnder extends TileChest implements IFacing, IOwned {
 
 	@Override
 	public EntityPlayer getOwner() {
-		return ModHelper.findPlayerByUUID(ownerUUID);
+		return PlayerUtils.findPlayerByUUID(ownerUUID);
 	}
 
 	@Override
