+++ src/main/java/T145/magistics/tiles/TileInfuser.java
@@ -1,13 +1,13 @@
 package T145.magistics.tiles;
 
-import T145.magistics.api.InventoryHelper;
 import T145.magistics.api.MagisticsApi;
 import T145.magistics.api.crafting.InfuserRecipe;
 import T145.magistics.api.tiles.IFacing;
 import T145.magistics.api.tiles.IOwned;
-import T145.magistics.api.tiles.TileMagisticsInventory;
 import T145.magistics.items.ItemShardDull;
-import T145.magistics.lib.ModHelper;
+import T145.magistics.lib.utils.InventoryHelper;
+import T145.magistics.lib.utils.PlayerUtils;
+import T145.magistics.tiles.core.TileMagisticsInventory;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.entity.player.EntityPlayer;
@@ -68,7 +68,7 @@ public class TileInfuser extends TileMagisticsInventory implements IFacing, IOwn
 
 	@Override
 	public EntityPlayer getOwner() {
-		return ModHelper.findPlayerByUUID(ownerUUID);
+		return PlayerUtils.findPlayerByUUID(ownerUUID);
 	}
 
 	@Override
